perf(call-center): lazy-load runtime and add smoke
